Fixing warnings; Qt4 apparently uses signed ints for QList.size().
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@8221 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/tools/xml/GenerateWindow.h
@@ -1423,7 +1423,7 @@ class WindowGeneratorAttribute : public GeneratorBase
     bool RequiresGetCurrentValues() const
     {
         int n = 0;
-        for (int i=0; i<fields.size(); i++)
+        for (size_t i=0; i<fields.size(); i++)
         {
             if (fields[i]->internal || !fields[i]->providesSourceGetCurrent())
                 continue;
